% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcfluutils-package.R
\docType{package}
\name{cdcfluutils}
\alias{cdcfluutils}
\alias{cdcfluutils-package}
\title{cdcfluutils - Utility functions for CDC flu forecasts for team Kernel of Truth}
\description{
Utility functions for generating and plotting forecasts from a variety of
   models used by team Kernel of Truth for the influenza forecasting
   challenge run by the CDC.
}
\author{
Evan Ray <eray@mtholyoke.edu>
}
