% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_e_matrix}
\alias{create_e_matrix}
\title{Create non-parametric matrix of backfill with dimensions
 n_seasons x n_weeks x n_lags}
\usage{
create_e_matrix()
}
\description{
Create non-parametric matrix of backfill with dimensions
 n_seasons x n_weeks x n_lags
}
