% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfill-utils.R
\name{create_ILI_revision_tables}
\alias{create_ILI_revision_tables}
\title{create tables of ILI trajectory revisions by epiweek}
\usage{
create_ILI_revision_tables()
}
\value{
list of n by length(observed_inc) matrix of samples for possible
revisions to ili, where n = (# regions)*(# past seasons) and
length(observed_inc) is dependent on the list component index
}
\description{
create tables of ILI trajectory revisions by epiweek
}
