% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-download-utils.R
\name{download_and_preprocess_hosp_data}
\alias{download_and_preprocess_hosp_data}
\title{Download and preprocess the latest CDC hospitalization data}
\usage{

  download_and_preprocess_hosp_data(latest_year = as.numeric(format(Sys.Date(),
  "\%Y")))
}
\arguments{
\item{latest_year}{year through which data should be downloaded, defaults to current year}
}
\value{
data frame with latest hospitalization data, preprocessed
}
\description{
Download and preprocess the latest CDC hospitalization data
}
