% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-download-utils.R
\name{download_backfill_data}
\alias{download_backfill_data}
\title{Download backfill data from the delphi api
and save to data folder.
Note: We may want to do this periodically to make sure we are up to date}
\usage{
download_backfill_data()
}
\description{
Download backfill data from the delphi api
and save to data folder.
Note: We may want to do this periodically to make sure we are up to date
}
