% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring-utils.R
\name{first_season_week_below_baseline}
\alias{first_season_week_below_baseline}
\title{Find the first season week incidence goes below baseline and doesn't go back
above it at the end of the season}
\usage{
first_season_week_below_baseline(location, season, incidence_data,
  target_variable, first_analysis_time_season_week = 10,
  last_analysis_time_season_week = cdcfluutils::get_num_MMWR_weeks_in_first_season_year(season)
  - 11)
}
\value{
integer season week, or NA if incidence never went above baseline
}
\description{
Find the first season week incidence goes below baseline and doesn't go back
above it at the end of the season
}
