% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_vars_regional}
\alias{historical_vars_regional}
\title{Variances of historical revisions to ilinet data}
\format{matrix}
\source{
Postprocessing from the Epidata API
}
\usage{
data(historical_vars_regional)
}
\description{
Variances of historical revisions to ilinet data
}
