% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde-utils.R
\name{make_kde_region_prediction}
\alias{make_kde_region_prediction}
\title{Make a KDE prediction for one region}
\usage{
make_kde_region_prediction(fit, cdc_region_string, inc_bins, inc_bin_names,
  season_has_EW53, season_week, n_sim)
}
\arguments{
\item{fit}{a kde_fit R object}

\item{cdc_region_string}{string for one CDC region}

\item{inc_bins}{list of incidence bins}

\item{inc_bin_names}{names of incidence bins}

\item{season_has_EW53}{indicator of whether the current season has a week 53}

\item{season_week}{string of the season week for which to make the prediction for}

\item{n_sim}{number of simulations to run}
}
\value{
data.frame of the prediction, in CDC format
}
\description{
Make a KDE prediction for one region
}
