% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx-utils.R
\name{merge_predx_samples}
\alias{merge_predx_samples}
\title{merge predx samples by location and target}
\usage{
merge_predx_samples(predx_df)
}
\arguments{
\item{predx_df}{a predx data frame with multiple rows that have samples at
the same location and target}
}
\value{
a single predx data frame with merged samples.  Only rows with
  predx_class Sample or SampleCat are retained.
}
\description{
merge predx samples by location and target
}
