% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-utils.R
\name{mmwr_week_to_season}
\alias{mmwr_week_to_season}
\title{get season in which an mmwr week falls}
\usage{
mmwr_week_to_season(mmwr_week, mmwr_year, first_season_week = 31)
}
\arguments{
\item{mmwr_week}{mmwr week as an integer (only the week)}

\item{mmwr_year}{year for mmwr week as an integer
first_season_week = 31,}

\item{first_season_week}{number of week in year corresponding to the first
week in the season.  For example, our code takes this value to be 31:
a new influenza season starts on the 31st week of each year.}
}
\value{
vector of the same length as mmwr_week with season in which the week
falls, in the format "2018/2019"
}
\description{
get season in which an mmwr week falls
}
