% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-utils.R
\name{mmwr_week_to_season_week}
\alias{mmwr_week_to_season_week}
\title{Convert mmwr week to season week}
\usage{
mmwr_week_to_season_week(mmwr_week, mmwr_year, first_season_week = 31)
}
\arguments{
\item{mmwr_week}{integer vector of weeks in year}

\item{mmwr_year}{either a single (four digit) integer year or a vector of
integer years with the same length as year_week}

\item{first_season_week}{number of week in year corresponding to the first
week in the season.  For example, our code takes this value to be 31:
a new influenza season starts on the 31st week of each year.}
}
\value{
vector of the same length as year_week with the week of the season
  that each observation falls on
}
\description{
Convert mmwr week to season week
}
