% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_reg_flu_data_with_backfill}
\alias{nat_reg_flu_data_with_backfill}
\title{National and Regional influenza incidence in the US (1997 - 2019) including all reports when available.}
\format{A data.frame with observations on weighted influenza-like illness 
measurements from all HHS regions and the national level.}
\source{
The cdcfluview R package.

The cdcfluview R package.
}
\usage{
data(nat_reg_flu_data_with_backfill)

data(nat_reg_flu_data_with_backfill)
}
\description{
A dataset of public influenza data from the US CDC.

A dataset of public influenza data from the US CDC.
}
