% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-utils.R
\name{pick_week}
\alias{pick_week}
\title{For each week specified by a vector of dates giving the week start dates,
determine whether the week contains a specified date.}
\usage{
pick_week(week_start_date, year_to_pick, month_to_pick, day_to_pick)
}
\arguments{
\item{week_start_date}{A vector of Date objects specifying the date of the first day in
the weeks of interest}

\item{year_to_pick}{integer or character giving the year to pick, e.g. "2010"}

\item{month_to_pick}{integer or character giving the month to pick, e.g. "12"}

\item{day_to_pick}{integer or character giving the day to pick, e.g. "22"}
}
\value{
a logical vector of the same length as time.  Entry i is TRUE if the
week beginning on week_start_date[i] contains the date specified by
year_to_pick, month_to_pick, and day_to_pick; FALSE otherwise.
}
\description{
For each week specified by a vector of dates giving the week start dates,
determine whether the week contains a specified date.
}
