% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx-utils.R
\name{predx_to_csv}
\alias{predx_to_csv}
\title{Combines predx format forecast files from multiple different runs and writes
the results to csv file.}
\usage{
predx_to_csv(file_path, ew, year)
}
\arguments{
\item{file_path}{file path to base directory for method.  should contain
subfolder predx with .rds files with predx format files and subfolder csv
into which results will be saved}

\item{ew}{epidemic week, character in format "01"}

\item{year}{year, numeric or character}
}
\value{
path to csv file
}
\description{
Combines predx format forecast files from multiple different runs and writes
the results to csv file.
}
