% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_and_augment_trajectory_samples}
\alias{preprocess_and_augment_trajectory_samples}
\title{Given a set of sampled trajectories for future disease incidence, preprocess
them in preparation for using them to generate forecast samples for
short-term and seasonal incidence targets.}
\usage{
preprocess_and_augment_trajectory_samples(trajectory_samples, round_digits,
  obs_data, prediction_target_var, first_season_obs_ind,
  analysis_time_ind = nrow(obs_data), analysis_time_season,
  analysis_time_epiweek, region, seasonal_target_week_limits,
  backfill_adjust = FALSE)
}
\arguments{
\item{trajectory_samples}{an n by h matrix of sampled incidence trajectories;
Each sampled trajectory is for the next h weeks. There are n trajectories}

\item{round_digits}{number of digits to which incidence will be rounded}

\item{obs_data}{a data frame containing observed incidence so far this
season; at a minimum, must contain two columns: season_week, indicating
the week of the current disease season, and the column given in
prediction_target_var}

\item{prediction_target_var}{character specifying column name in obs_data for
which we are generating predictions.}

\item{first_season_obs_ind}{index of row in obs_data for the first week of
the season for which we are generating predictions.}

\item{analysis_time_ind}{index of row in obs_data that contains the most
recent observation of incidence (default nrow(obs_data))}

\item{analysis_time_season}{in format 2018/2019}

\item{analysis_time_epiweek}{in format 201840}

\item{seasonal_target_week_limits}{vector of length 2 specifying season 
 weeks that will be used for identifying seasonal targets such as peak
incidence; sampled or observed incidence outside of this range will be
set to NA.}

\item{backfill_adjust}{logical; adjust observed data for backfill?}
}
\value{
an n by W matrix of preprocessed trajectory samples, where
    W = (# of weeks observed so far this season) + h
}
\description{
Given a set of sampled trajectories for future disease incidence, preprocess
them in preparation for using them to generate forecast samples for
short-term and seasonal incidence targets.
}
