% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPthingy-utils.R
\name{preprocess_reg_data_for_GPthingy_stan}
\alias{preprocess_reg_data_for_GPthingy_stan}
\title{Preprocess state level flu data and set up a data object to pass to stan}
\usage{
preprocess_reg_data_for_GPthingy_stan(us_flu, seasons, predict_season,
  predict_last_obs_season_week, return_extras = FALSE,
  knot_limits = c(1, 53L), DX = 10, get_all_region_seasons = FALSE,
  pred_all_t = FALSE)
}
\arguments{
\item{us_flu}{data frame of regional and national flu observations}

\item{seasons}{character vector of seasons to include in "2014/2015" format}

\item{predict_season}{which season are we generating predictions for}

\item{predict_last_obs_season_week}{integer season week from which to predict}

\item{return_extras}{boolean, default FALSE: return extra things not needed for
stan, but that might be helpful for other things (like making plots)}
}
\value{
list object with data for stan
}
\description{
Preprocess state level flu data and set up a data object to pass to stan
}
