% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-utils.R
\name{season_week_to_mmwr_year}
\alias{season_week_to_mmwr_year}
\title{return mmwr year corresponding to a given season week}
\usage{
season_week_to_mmwr_year(season_week, season, first_season_week = 31)
}
\arguments{
\item{season_week}{season week as integer from 1 to 53}

\item{season}{in the format "2018/2019"}

\item{first_season_week}{integer specifying the first mmwr week of the season}
}
\description{
return mmwr year corresponding to a given season week
}
