% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_flu_data}
\alias{state_flu_data}
\title{State-level influenza incidence in the US (1997 - 2018)}
\format{A data.frame with 22,310 observations on influenza-like illness 
measurements from all US states.}
\source{
The cdcfluview R package.
}
\usage{
data(state_flu_data)
}
\description{
A dataset of public influenza data from the US CDC.
}
